requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
pandas
openpyxl
numpy
Werkzeug
python-dotenv
PyJWT